<fmt:formatNumber>
Problem:
-------

  <fmt:formatNumber value="$books.rows[0].columns[4]" type="currency"/>

throws exception, since supplied value is instance of Column, but
<fmt:formatNumber> requires it to be a String or Number.

Workaround:
----------
If 'value' is neither a Number nor a String, it is converted to a
String.

Note:
----
It seems like we need to add methods to the Row interface that
will return the column "value" as opposed to an instance of "Column".

Otherwise, any tag that requires a special input type will not be able
to process the result of a database lookup.

In the case of <fmt:formatNumber>, this is not such a big deal, as
this action already accepts String objects (in addition to instances
of java.lang.Number), but this may not be true for other tags.

We will raise this issue with the expert group.

(Similar workaround implemented for <fmt:formatDate>.)


git-svn-id: https://svn.apache.org/repos/asf/jakarta/taglibs/proper/standard/trunk@217669 13f79535-47bb-0310-9956-ffa450edef68