Removed 'parseLocale' attribute from <fmt:formatDate> and <fmt:formatNumber>.
If the value to be formatted is given as a String, it is now always
parsed using the "en" locale.

If the String is formatted with a different locale than "en", it must
first be parsed to a java.util.Date or java.lang.Number with
<fmt:parseDate> or <fmt:parseNumber>, and the result can then be used
with <fmt:formatDate> or <fmt:formatNumber>, respectively.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/taglibs/proper/standard/trunk@217713 13f79535-47bb-0310-9956-ffa450edef68