Fixed Bug 7158 - EL exception message should contain attribute name/value
Made a few general changes to clean up the formatting of EL exception
messages.

Changed the token definitions that have multiple forms, like "<" and
"lt" - broke each of those out as separate tokens so that the error
message would print "<", "lt" as expected tokens, not "<LE>".

Customized the error message of the ParseException so that:

  * the attribute name and value are included in the error message
  * the line/column numbers are no longer in the error message
  * the list of expected tokens is put on a single line, not broken
    out onto separate lines

This was done without actually changing the definition of
ParseException (in case JavaCC is run again and overwrites any
customizations).

Also added a catch-all token so that any input can be tokenized, hence
no TokenMgrErrors - all errors now show up as ParseExceptions,
formatted as described above.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/taglibs/proper/standard/trunk@217797 13f79535-47bb-0310-9956-ffa450edef68