Bug 7158 - EL exception message should contain attribute name/value
The previous fix only covered parse exceptions (which would be thrown
either at runtime or translation time), but didn't cover the purely
runtime ELExceptions.  Those are now being caught and wrapped with a
message indicating the attribute name and value.

I don't know if anyone cares, but this might be a slight security
problem since it's possible for some of a page's "source code" (i.e.,
the attribute value expression) to be visible to the end user if an
exception makes it out that far.  But I suppose exceptions in general
can expose a lot of undesirable information to the end user, so this
is probably no more of a problem than it has always been.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/taglibs/proper/standard/trunk@217813 13f79535-47bb-0310-9956-ffa450edef68