Bug 9257 - ELParser doesn't parse non-ascii value
As Kan suggested, I've added UNICODE_INPUT = true to the ELParser.jj
options.  This doesn't seem to have caused any regression in the
existing command line parser tests.

I also added a new command line parser test to try inputting a
non-ascii string.  I've confirmed that without the fix, the string
causes an error, whereas with the fix, the string is processed
correctly.

So, this appears to be fixed.  However, I don't have easy access to
non-ascii pages so I haven't completely confirmed this fix in an
actual JSP.  I would appreciate it if someone would confirm this fix
(Kan?)


git-svn-id: https://svn.apache.org/repos/asf/jakarta/taglibs/proper/standard/trunk@218014 13f79535-47bb-0310-9956-ffa450edef68