Bugzilla #31313 http://nagoya.apache.org/bugzilla/show_bug.cgi?id=31313
According to the W3C document the schemaLocation declaration is made up of a pair of URI references. One defines the namespace name and the other is a hint as to the location of the schema document. The space in the schemaLocation definition separates the two declarations.

For a complete explanation you can take a look at part 3 of:

http://www.w3.org/TR/xmlschema-1/#schema-loc

In the JSTL case the location is not an absolute URI which can be confusing especially to some IDE's. I went ahead and changed it to be an absoluate URI pointing to the schema document location.

Note that there is a space character between the namespace declaration and schema document location. That space character is intentional.


git-svn-id: https://svn.apache.org/repos/asf/jakarta/taglibs/proper/standard/trunk@218346 13f79535-47bb-0310-9956-ffa450edef68