Applying Jeremy Boynes' refactoring of SetSupport per #49548
"Patch primarily refactors the large "if" clause that sets
the result into smaller methods. It also clears up IDE warnings (at least from
IDEA).

One side effect was I18N of the exception message for "SET_BAD_DEFERRED_SCOPE"

I18N also needed access to the original value (which was not being retained)
and in conjunction with that I replaced the "scopeSpecified" indicator with a
check on whether the "scope" attribute is null.

I also changed the comment where the exception is thrown when target is null.
This can happen as target is a request-time attribute and the spec does define
this behaviour.

When setting a bean property, the for loop now exits after setting its property
rather than continuing to iterate through all of them. This should have no
impact unless a bean can have two PropertyDescriptors with the same name and I
don't believe it can.

Finally, all throws of JspException were replaced with JspTagException as these
are errors coming from a Tag Handler rather than the JSP Engine itself."



git-svn-id: https://svn.apache.org/repos/asf/tomcat/taglibs/standard/trunk@960817 13f79535-47bb-0310-9956-ffa450edef68