Fix compilation failure in a test class when using JDK 1.5. The XPathFactory.newInstance(String, String, ClassLoader) method is @since Java 1.6.
From code review, the only difference between old and new methods is presence of the second argument, "com.sun.org.apache.xpath.internal.jaxp.XPathFactoryImpl". The null class loader is interpreted as TCCL by both methods, so there is no difference.
This test is @Ignore'd so it is not really tested.

git-svn-id: https://svn.apache.org/repos/asf/tomcat/taglibs/standard/trunk@1657653 13f79535-47bb-0310-9956-ffa450edef68